comment out amt device id for tetsing
