+++ app/portainer/components/datatables/amt-devices-datatable/amtDevicesDatatable.html
@@ -16,13 +16,13 @@
         <td>
           <div class="btn-group btn-group-xs" role="group" aria-label="..." style="display: inline-flex;">
             <a style="margin: 0 2.5px;" title="Power up the device">
-              <i class="fa fa-file-alt space-right" aria-hidden="true"></i>
+              <i class="fa fa-plug space-right" aria-hidden="true"></i>
             </a>
             <a style="margin: 0 2.5px;" title="Power off the device">
-              <i class="fa fa-file-alt space-right" aria-hidden="true"></i>
+              <i class="fa fa-power-off space-right" aria-hidden="true"></i>
             </a>
             <a style="margin: 0 2.5px;" title="Restart the device">
-              <i class="fa fa-file-alt space-right" aria-hidden="true"></i>
+              <i class="fa fa-sync space-right" aria-hidden="true"></i>
             </a>
           </div>
         </td>
